+++ src/libcpp/cstdlib.h
@@ -44,7 +44,7 @@ void merge_sort(int array[], int first, int last);
 void bit_set(uint16_t* address, uint16_t bit_number);
 bool bit_clear(uint16_t* address, uint16_t bit_number);
 void memory_manager_init(void);
-void* malloc(uint16_t size);
+void* malloc(uint32_t size);
 void* realloc(void* ptr, uint32_t size_old, uint32_t size_new);
 void* free(void* ptr);
 void wait_key(uint8_t key);
